requirements.txt
Django==4.0.5
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
drf-yasg==1.20.0
Pillow==9.1.1
psycopg2-binary==2.9.3
